+++ packages/quarks.r3f/src/components/index.ts
@@ -1,0 +1,2 @@
+export { ParticleSystem } from './ParticleSystem';
+export { QuarksEffect } from './QuarksEffect';